Don't rely on reqwest::blocking::get for HTTP requests
