+++ nemo-physical/Cargo.toml
@@ -27,6 +27,7 @@ linked-hash-map = "0.5.6"
 howlong = { version = "0.1", optional = true }
 rio_turtle = "0.8.4"
 rio_xml = "0.8.4"
+reqwest = "0.11.18"
 
 [dev-dependencies]
 arbitrary = { version = "1", features = ["derive"] }
