+++ nemo/Cargo.toml
@@ -10,18 +10,16 @@ readme = "README.md"
 repository.workspace = true
 
 [features]
-# TODO: Add "http-resource-provider"
 default = ["timing"]
 # Allows building for web assembly environments
 # Enables the "js" feature of the "getrandom" crate
 # This feature cannot be used together with the "timing" feature, because the "howlong" crate does not support web assembly environments
-# This feature cannot be used together with the "http-resource-provider" feature, because the "reqwest" crate does not support web assembly environments with the "blocking" feature
 js = ["getrandom/js"]
-http-resource-provider = ["dep:reqwest"]
 no-prefixed-string-dictionary = ["nemo-physical/no-prefixed-string-dictionary"]
 timing = ["nemo-physical/timing"]
 
 [dependencies]
+nemo-physical = { path = "../nemo-physical", default-features = false }
 log = "0.4"
 nom = "7.1.1"
 macros = { path = "../libs/macros" }
@@ -32,15 +30,14 @@ thiserror = "1.0"
 flate2 = "1"
 sanitise-file-name = "1.0.0"
 nom_locate = { version = "4.1.0", features = [ "runtime-dispatch-simd" ] }
-getrandom = {version = "0.2.9", default-features = false}
-reqwest = { version = "0.11.18", features = ["blocking"], optional = true }
-
-nemo-physical = { path = "../nemo-physical", default-features = false }
+getrandom = { version = "0.2.9", default-features = false }
 path-slash = "0.2.1"
 rio_api = "0.8.4"
 rio_turtle = "0.8.4"
 rio_xml = "0.8.4"
 oxiri = "0.2.2"
+tokio = { version = "1.29.1", features = [ "rt" ] }
+reqwest = { version = "0.11.18" }
 
 [dev-dependencies]
 env_logger = "*"
